Add files via upload
I spent the day today really looking into the Exploratory Data Analysis(EDA) and came up with some ideas that I think we could look further into. 

As you all probably know, data can be classified in many ways (categorical vs. numerical, qualitative vs. quantitative, etc.) For conducting an EDA, I was thinking that we should separate out the data as qualitative vs. quantitative. It is important to note that from a brief glance at the data it might look as if most of the data is numerical in nature(either having a date type of int, or num) but that does not necessarily mean that the data is quantitative. For example, a person identification number may look as if its numerical, but it cannot be quantified and therefore is qualitative. 

I soon realized that most of the data is qualitative in nature, and that there are very few quantitative variables (possibly backward_cities, and forward_cities). So, I decided to create an additional quantitative variable called years_elapsed, which takes the difference in years between the application date of the patent and the grant date.

I factored the variables that are BOTH qualitative and categorical in nature and created new columns for those variables that way they are easier to compare. E.g. patent_type_factor

I exported a new csv called "patentdatarevised.csv" where I moved all the factors to the left side of the data frame, followed by the years_elapsed, backward_cities, and forward_cities. That way all the useful data is next to each other. I did not delete the original data, I left it in the data frame for reference but pushed it to the right side of the data frame. 

Here are some ideas I have for the EDA:

When working with the qualitative data (anything with _factor in the variable name), we can create contingency tables to look at the variables that we are interested in. For example, if we want to look at the relationship between the patent type, and the nationality of the company with the patent we could do:
table(mydata$patent_type_factor, mydata$company_nationality_factor)

or if we wanted to look at the relationship between the patent type and assignee type, we could do:
table(mydata$patent_type_factor, mydata$assignee_type_factor)

For the numerical data, we can work with the years_elapsed, and possibly the backward_cities, and forward_cities. We can use ggplot and create dotplots, boxplots, histograms, etc. Not sure really how to approach this, but maybe some of you have some ideas.

Here I have attached the new and revised data set. I hope this is helpful for you all.